# [CUDA] Split/Concat Kernel Optimization