--- a/onnxruntime/core/providers/cuda/tensor/concat.cc
+++ b/onnxruntime/core/providers/cuda/tensor/concat.cc
@@ -1,8 +1,9 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-#include "concat.h"
-#include "concat_impl.h"
+#include "core/providers/cuda/tensor/concat.h"
+
+#include "core/providers/cuda/tensor/concat_impl.h"
 
 namespace onnxruntime {
 namespace cuda {
@@ -63,31 +64,41 @@ Status Concat::ComputeInternal(OpKernelContext* ctx) const {
       axis_dimension_input_output_mapping.at(index++) = i;
     }
   }
-  std::vector<int64_t> concat_sizes_range(concat_sizes);
-  for (size_t i = 1; i < concat_sizes_range.size(); ++i) {
-    concat_sizes_range[i] += concat_sizes_range[i - 1];
-  }
 
-  CudaAsyncBuffer<int64_t> concat_sizes_gpu(this, concat_sizes);
-  CudaAsyncBuffer<int64_t> axis_dimension_input_output_mapping_gpu(this, axis_dimension_input_output_mapping);
-  CudaAsyncBuffer<int64_t> concat_sizes_range_gpu(this, concat_sizes_range);
-  ORT_RETURN_IF_ERROR(concat_sizes_gpu.CopyToGpu());
-  ORT_RETURN_IF_ERROR(axis_dimension_input_output_mapping_gpu.CopyToGpu());
-  ORT_RETURN_IF_ERROR(concat_sizes_range_gpu.CopyToGpu());
-  ORT_RETURN_IF_ERROR(input_ptr.CopyToGpu());
+  auto element_bytes = p.output_tensor->DataType()->Size();
   int block_size_inside_axis_dim = static_cast<int>(p.output_axis_pitch / p.output_tensor->Shape()[p.axis]);
   int block_size_including_axis_dim = static_cast<int>(p.output_axis_pitch);
-  auto element_bytes = p.output_tensor->DataType()->Size();
-  ORT_RETURN_IF_ERROR(ConcatImpl(Stream(),
-                                 element_bytes,
-                                 block_size_including_axis_dim,
-                                 block_size_inside_axis_dim,
-                                 concat_sizes_gpu.GpuPtr(),
-                                 concat_sizes_range_gpu.GpuPtr(),
-                                 axis_dimension_input_output_mapping_gpu.GpuPtr(),
-                                 p.output_tensor->MutableDataRaw(),
-                                 input_ptr.GpuPtr(),
-                                 p.output_num_elements));
+  if (std::all_of(concat_sizes.begin(), concat_sizes.end(), [&](int64_t size) { return size == concat_sizes[0]; })) {
+    if (input_count <= 32) {
+      TArray<const void*, 32> input_ptr_array(input_count);
+      for (int i = 0; i < input_count; ++i) input_ptr_array[i] = input_ptr_cpuspan[i];
+      ORT_RETURN_IF_ERROR(ConcatSameConcatDimImpl(
+          Stream(), element_bytes, block_size_including_axis_dim, block_size_inside_axis_dim, concat_sizes[0],
+          p.output_tensor->MutableDataRaw(), input_ptr_array, static_cast<size_t>(p.output_num_elements)));
+    } else {
+      ORT_RETURN_IF_ERROR(input_ptr.CopyToGpu());
+      ORT_RETURN_IF_ERROR(ConcatSameConcatDimImpl(
+          Stream(), element_bytes, block_size_including_axis_dim, block_size_inside_axis_dim, concat_sizes[0],
+          p.output_tensor->MutableDataRaw(), input_ptr.GpuPtr(), static_cast<size_t>(p.output_num_elements)));
+    }
+  } else {
+    CudaAsyncBuffer<int64_t> concat_sizes_gpu(this, concat_sizes);
+    CudaAsyncBuffer<int64_t> axis_dimension_input_output_mapping_gpu(this, axis_dimension_input_output_mapping);
+    std::vector<int64_t> concat_sizes_range(concat_sizes);
+    for (size_t i = 1; i < concat_sizes_range.size(); ++i) {
+      concat_sizes_range[i] += concat_sizes_range[i - 1];
+    }
+    CudaAsyncBuffer<int64_t> concat_sizes_range_gpu(this, concat_sizes_range);
+    ORT_RETURN_IF_ERROR(concat_sizes_gpu.CopyToGpu());
+    ORT_RETURN_IF_ERROR(axis_dimension_input_output_mapping_gpu.CopyToGpu());
+    ORT_RETURN_IF_ERROR(concat_sizes_range_gpu.CopyToGpu());
+    ORT_RETURN_IF_ERROR(input_ptr.CopyToGpu());
+    ORT_RETURN_IF_ERROR(ConcatImpl(Stream(), element_bytes, block_size_including_axis_dim, block_size_inside_axis_dim,
+                                   concat_sizes_gpu.GpuPtr(), concat_sizes_range_gpu.GpuPtr(),
+                                   axis_dimension_input_output_mapping_gpu.GpuPtr(), p.output_tensor->MutableDataRaw(),
+                                   input_ptr.GpuPtr(), static_cast<size_t>(p.output_num_elements)));
+  }
+
   return Status::OK();
 }
 }  // namespace cuda