--- a/onnxruntime/core/providers/cuda/tensor/concat_impl.h
+++ b/onnxruntime/core/providers/cuda/tensor/concat_impl.h
@@ -9,26 +9,15 @@
 namespace onnxruntime {
 namespace cuda {
 
-template<typename InputIndexToMemoryMap>
-Status ConcatSameConcatDimImpl(cudaStream_t stream,
-                  const size_t element_bytes,
-                  const int block_size_including_axis_dim,
-                  const int block_size_inside_axis_dim,
-                  const int64_t concat_size,
-                  void* output_data,
-                  const InputIndexToMemoryMap input_ptr,
-                  const size_t N);
+template <typename InputDataArray>
+Status ConcatSameConcatDimImpl(cudaStream_t stream, const size_t element_bytes, const int block_size_including_axis_dim,
+                               const int block_size_inside_axis_dim, const int64_t concat_size, void* output_data,
+                               const InputDataArray input_data, const size_t output_size);
 
-Status ConcatImpl(cudaStream_t stream,
-                  const size_t element_bytes,
-                  const int block_size_including_axis_dim,
-                  const int block_size_inside_axis_dim,
-                  const int64_t* concat_sizes,
-                  const int64_t* concat_sizes_range,
-                  const int64_t* axis_dimension_input_output_mapping,
-                  void* output_data,
-                  const void** input_ptr,
-                  const size_t N);
+Status ConcatImpl(cudaStream_t stream, const size_t element_bytes, const int block_size_including_axis_dim,
+                  const int block_size_inside_axis_dim, const int64_t* concat_sizes, const int64_t* concat_sizes_range,
+                  const int64_t* axis_dimension_input_output_mapping, void* output_data, const void** input_data,
+                  const size_t output_size);
 
 }  // namespace cuda
 }  // namespace onnxruntime