--- a/onnxruntime/core/providers/cuda/tensor/split_impl.h
+++ b/onnxruntime/core/providers/cuda/tensor/split_impl.h
@@ -9,29 +9,15 @@
 namespace onnxruntime {
 namespace cuda {
 
-template<typename OutputIndexToMemoryMap>
-Status SplitSameSplitDimImpl(cudaStream_t stream,
-                 const size_t element_size,
-                 const int block_size_including_axis_dim,
-                 const int block_size_inside_axis_dim,
-                 const int64_t split_size,
-                 const int num_outputs,
-                 const void* input_data,
-                 OutputIndexToMemoryMap output_ptr,
-                 const size_t N);
+template <typename OutputDataArray>
+Status SplitSameSplitDimImpl(cudaStream_t stream, const size_t element_size, const int block_size_including_axis_dim,
+                             const int block_size_inside_axis_dim, const int64_t split_size, const int num_outputs,
+                             const void* input_data, OutputDataArray output_data, const size_t input_size);
 
-
-Status SplitImpl(cudaStream_t stream,
-                 const size_t element_size,
-                 const int block_size_including_axis_dim,
-                 const int block_size_inside_axis_dim,
-                 const int64_t* split_sizes,
-                 const int64_t* split_sizes_range,
-                 const int64_t* axis_dimension_input_output_mapping,
-                 const int num_outputs,
-                 const void* input_data,
-                 void** output_ptr,
-                 const size_t N);
+Status SplitImpl(cudaStream_t stream, const size_t element_size, const int block_size_including_axis_dim,
+                 const int block_size_inside_axis_dim, const int64_t* split_sizes, const int64_t* split_sizes_range,
+                 const int64_t* axis_dimension_input_output_mapping, const int num_outputs, const void* input_data,
+                 void** output_data, const size_t input_size);
 
 }  // namespace cuda
 }  // namespace onnxruntime